Report NETWM_CHANGED_WINDOW_ICON when icon gets changed via _NET_WM_ICON message. amd64 requires '-fPIC' even if library is compiled as static. Some warning reduction.

+++ examples/efile.cpp
@@ -33,7 +33,7 @@ int main(int argc, char** argv) {
 
 		//printf("%s: %30s (%s)\n", argv[i], mt.comment().c_str(), mt.type().c_str());
 		printf("%s:", argv[i]);
-		printf("%*s ", mlen - strlen(argv[i]), "");
+		printf("%*s ", (int)(mlen - strlen(argv[i])), "");
 		printf("%s (%s)\n", mt.comment().c_str(), mt.type().c_str());
 	}
